requirements.txt
beautifulsoup4==4.12.3
httpx
requests
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client